requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.0            # via setuptools
apscheduler==3.3.1
click==6.7                # via clickclick, flask
clickclick==1.2.1         # via connexion
connexion==1.1.5
Flask-Script==2.0.5
Flask==0.12               # via connexion, flask-script
itsdangerous==0.24        # via flask
Jinja2==2.9.5             # via flask
jsonschema==2.6.0         # via connexion, swagger-spec-validator
MarkupSafe==0.23          # via jinja2
packaging==16.8           # via setuptools
pyparsing==2.1.10         # via packaging
pytimeparse==1.1.6
pytz==2016.10             # via apscheduler, tzlocal
PyYAML==3.12              # via clickclick, connexion
requests==2.13.0          # via connexion
six==1.10.0               # via apscheduler, connexion, packaging, setuptools, swagger-spec-validator
strict-rfc3339==0.7       # via connexion
swagger-spec-validator==2.0.2  # via connexion
tzlocal==1.3              # via apscheduler
uwsgi==2.0.14
Werkzeug==0.11.15         # via flask

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via apscheduler, swagger-spec-validator
